Combined the functionality of the Combination and Permutation test suites into the Combination test suite.
The Combination test suite can now use a constructor to pass
parameters to the test or use fields or methods to pass the parameters
to the test class.  The Attributes annotation has been enhanced to
include an attributes field that is an array of string.  This array
is assigned with the names of the attributes that have associated
public field or method access.  If this empty, which is the default
setting, then the Combination test suite will assume that there is a
constructor that has the correct number of parameters and parameter
types.

Prior to this the Combination test suite expected a Map as the
values of the attributes but this has changed and it now expects
a list that contains object arrays for each attribute.  The
map key were used to define the attribute names but now this is
accomplished using the Attributes annotation attributes field.

The CombinationTestTest has been updated to include the tests that
were part of the PermutationTestTest test since the Combination
test suite includes the same feature.

The Permutation test suite has been removed along with the
PermutationTestTest test.